Implement MainContent component to use drawer context in, make drawer fixed and push main content instead of overlay
