+++ client/src/App.tsx
@@ -5,7 +5,7 @@ import "@fontsource/roboto/500.css";
 import "@fontsource/roboto/700.css";
 import { createBrowserRouter, RouterProvider } from "react-router-dom";
 import HomePage from "./pages/Home/Home";
-import PrivateLayout from "./layouts/Private";
+import PrivateLayout from "./layouts/Private/Private";
 import PlaygroundPage from "./pages/Playground/Playground";
 import LoginPage from "./pages/Login/Login";
 import MainLayout from "./layouts/Main";
